fetch client data from server
