+++ web/src/app/authorize/loading.tsx
@@ -1,0 +1,5 @@
+import { Container } from '@mui/material';
+
+export default function Loading() {
+    return <Container>loading...</Container>;
+}
